requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
Flask==3.0.2
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.21
PyJWT==2.8.0
Werkzeug==3.0.1
WTForms==3.1.2
